linkedin/job_posts/requirements.txt
numpy
pandas
selenium
bs4
lxml
azure-storage-blob